fix pools type casting bug

+++ types/assets.go
@@ -49,7 +49,7 @@ func (s *Asset) Fetch(algod *algod.Client) error {
 }
 
 //TODO: is call and hash methods ok?
-func (s *Asset) Call(amount float64) AssetAmount {
+func (s *Asset) Call(amount uint64) AssetAmount {
 	return AssetAmount{*s, amount}
 }
 
@@ -63,12 +63,12 @@ func (s *Asset) String() string {
 
 type AssetAmount struct {
 	Asset  Asset
-	Amount float64
+	Amount uint64
 }
 
 func (s *AssetAmount) Mul(o float64) AssetAmount {
 
-	return AssetAmount{s.Asset, s.Amount * o}
+	return AssetAmount{s.Asset, uint64(float64(s.Amount) * o)}
 }
 
 func (s *AssetAmount) Add(o AssetAmount) (AssetAmount, error) {
